style: 컴포넌트 스타일, 로직 파일 분리
