+++ src/components/DateRange/styles.ts
@@ -1,0 +1,36 @@
+import styled from 'styled-components';
+
+export const DeadlineField = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+
+  .deadline-label {
+    display: flex;
+    gap: 0.5rem;
+    align-items: center;
+
+    .label-name {
+      color: #76808e;
+    }
+  }
+
+  .deadline-prop {
+    padding-left: 0.5rem;
+    display: flex;
+    gap: 1rem;
+    align-items: center;
+    justify-content: center;
+
+    .prop-name-container {
+      display: flex;
+      gap: 0.5rem;
+      align-items: center;
+      justify-content: center;
+
+      .prop-name {
+        font-size: 0.8rem;
+      }
+    }
+  }
+`;
